Combine Rcpp fns into one file to compile in 4.0
